debt: remove previous homepage
